server/requirements.txt
anyio==3.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
future==0.18.3
h11==0.14.0
huggingface-hub==0.12.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
more-itertools==9.0.0
mutagen==1.46.0
nltk==3.8.1
numpy==1.24.2
openai-whisper==20230124
packaging==23.0
Pillow==9.4.0
pycryptodomex==3.17
pydantic==1.10.5
pytube==12.1.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sniffio==1.3.0
starlette==0.25.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.2.3
yt-dlp==2023.2.17
